requirements.txt
aiodns==2.0.0
aiofile==1.5.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cchardet==2.1.6
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptofeed==1.4.1
Flask==1.1.2
idna==2.9
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
numpy==1.18.5
pandas==1.0.4
playsound==1.2.2
PyAudio==0.2.11
pycares==3.1.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
redis==3.5.3
requests==2.23.0
selenium==3.141.0
six==1.15.0
sortedcontainers==2.2.2
SpeechRecognition==3.8.1
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
yapic.json==1.5.1
yarl==1.4.2
